Big Changes: App is no longer only for testing.
- The app now works normally theoretically (next comes bug detecting, but up until now the app looks good for non-abnormal usage).
- Now each newly installed version of the app will have a unique slot in Firebase.
- Multiple users can work simultaneously now with no problem found, unlike before where the app can barely be used for testing on 1 device and had to be re-installed every time we want to use it.
Next step is implementing the deletion of plants functionality and the removal of the user and his/her data from Firebase every time the app is uninstalled. Bug detecting comes next, then we can safely work on the UI and the 'friendly' features we wanna add as their bugs are solely constrained to their own usage, they will not ruin any other part of the app.